+++ src/IPhreeqcPhast/IPhreeqc/phreeqcpp/utilities.cpp
@@ -1527,7 +1527,7 @@ status(int count, const char *str, bool rk_string)
 
 #ifndef NPP
 	t2 = clock();
-	if ((int) (1e3 / CLOCKS_PER_SEC * (t2 - status_timer)) > status_interval)
+	if ((int) (1e3 / (double)CLOCKS_PER_SEC * ((double)t2 -(double)status_timer)) > status_interval)
 #endif
 	{
 		status_timer = t2;
